auth_krb5: Remove vestigial krbtf call
auth_krb5 is not using a ccache at all any more; even if it were,
use of `MEMORY:0` was no longer conditional so this was effectively
a hardcoded string.

Because of that hardcoding krbtf itself could no longer be described
as a shared implementation; it had one code path that returned krb5's
magic string and another that handled krb4.