Fix enabling SASL_SEC_NONSTD_CBIND
This security flag is supposed to allow explicitly enabling channel
bindings for the GSSAPI mechanism. However, if I set this flag in the
test programs it results in the mechanism being rejected when the
library checks to see if it's suitable. Adding it to the flags defined
in the plugins fixed the tests, but also appeared to revert to channel
bindings being enabled by default.

The only way I was able to get the tests working and still require
explicit enablement was to ignore this flag while doing the
comparisons, which feels wrong but it's been over a year and no one
has offered an alternative solution.

Signed-off-by: Paul Arthur <paul.arthur@flowerysong.com>